fix(switch): resolve symlinked paths to existing worktrees
`worktree_at_path` compared paths via lexical normalization, so a
symlink whose target was a worktree did not match — causing
`wt switch /tmp/link-to-worktree` to fail with `No branch named ...`.

Move `paths_match` / `canonicalize_with_parents` from
`commands/worktree/resolve.rs` into `worktrunk::path` and switch
`worktree_at_path` to the symlink-aware comparison.

Closes #2460
